Update imports after reorganization
